# src/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "2팀-How Do I Look",
    "description": "이 프로젝트는 코드잇 Node.js 백엔드 스프린트의 초급 팀 프로젝트입니다.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/tags/": {
      "get": {
        "tags": ["태그"],
        "summary": "태그 목록 조회",
        "description": "스타일과 연결된 태그 중 인기 태그 10개를 조회합니다.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "tagname": {
                        "type": "string",
                        "example": "casual"
                      },
                      "count": {
                        "type": "integer",
                        "example": 120
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["태그"],
        "summary": "태그 생성",
        "description": "새로운 태그를 시스템에 추가합니다. 태그 이름을 제공하여 시스템에 등록합니다.",
        "requestBody": {
          "required": true,
          "description": "생성할 태그의 이름",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tagname": {
                    "type": "string",
                    "description": "새로운 태그의 이름",
                    "example": "tech"
                  }
                },
                "required": ["tagname"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tagname": {
                      "type": "string",
                      "example": "tech"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict (이미 존재하는 태그의 경우)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "이미 존재하는 태그입니다."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tags/{tagname}": {
      "get": {
        "tags": ["태그"],
        "summary": "특정 태그의 스타일 조회",
        "description": "지정된 태그명을 포함하는 스타일 목록을 조회합니다.",
        "parameters": [
          {
            "name": "tagname",
            "in": "path",
            "required": true,
            "description": "조회할 태그의 이름",
            "schema": {
              "type": "string",
              "example": "casual"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer", "example": 0 },
                      "title": {
                        "type": "string",
                        "example": "캐주얼 데일리룩"
                      },
                      "nickname": {
                        "type": "string",
                        "example": "fashionista01"
                      },
                      "content": {
                        "type": "string",
                        "example": "편안한 캐주얼 스타일이에요!"
                      },
                      "imageUrls": {
                        "type": "array",
                        "items": { "type": "string", "format": "uri" },
                        "example": [
                          "https://example.com/style1-1.jpg",
                          "https://example.com/style1-2.jpg"
                        ]
                      },
                      "viewCount": { "type": "integer", "example": 120 },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-01T12:00:00.000Z"
                      },
                      "categories": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "integer", "example": 28 },
                            "type": { "type": "string", "example": "TOP" },
                            "name": {
                              "type": "string",
                              "example": "화이트 오버핏 티셔츠"
                            },
                            "brand": { "type": "string", "example": "Uniqlo" },
                            "price": { "type": "number", "example": 29.99 },
                            "styleId": { "type": "integer", "example": 0 }
                          }
                        }
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "tagname": { "type": "string", "example": "casual" }
                          }
                        }
                      },
                      "curations": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "integer", "example": 0 },
                            "nickname": {
                              "type": "string",
                              "example": "닉네임"
                            },
                            "content": {
                              "type": "string",
                              "example": "string"
                            },
                            "trendy": { "type": "integer", "example": 10 },
                            "personality": { "type": "integer", "example": 0 },
                            "practicality": { "type": "integer", "example": 0 },
                            "costEffectiveness": {
                              "type": "integer",
                              "example": 0
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time",
                              "example": "2024-02-02T13:00:00.000Z"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (태그명이 올바르지 않거나 존재하지 않음)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid tag name"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["태그"],
        "summary": "태그 삭제",
        "description": "특정 태그를 삭제합니다. 단, 사용 중인 태그는 삭제할 수 없습니다.",
        "parameters": [
          {
            "name": "tagname",
            "in": "path",
            "required": true,
            "description": "삭제할 태그의 이름",
            "schema": {
              "type": "string"
            },
            "example": "casual"
          }
        ],
        "responses": {
          "204": {
            "description": "태그 삭제 성공 (No Content)"
          },
          "400": {
            "description": "현재 사용 중인 태그는 삭제할 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "현재 사용 중인 태그는 삭제할 수 없습니다."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 태그",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "존재하지 않는 태그입니다."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}